+++ nbodykit/io/stack.py
@@ -3,6 +3,7 @@ from . import tools
 from six import string_types
 import numpy
 import os
+from glob import glob
 import inspect
 
 class FileStack(FileType):
@@ -20,7 +21,7 @@ class FileStack(FileType):
         the type of file class to initialize
     path : str
         list of file names, or string specifying single file or
-        containing a glob-like '*' pattern
+        a glob pattern.
     *args :
         additional arguments to pass to the ``filetype`` instance
         during initialization
@@ -40,21 +41,17 @@ class FileStack(FileType):
         if isinstance(path, list):
             filenames = path
         elif isinstance(path, string_types):
-            if '*' in path:
-                from glob import glob
-                filenames = list(map(os.path.abspath, sorted(glob(path))))
-            else:
-                if not os.path.exists(path):
-                    raise FileNotFoundError(path)
-                filenames = [os.path.abspath(path)]
+            filenames = list(map(os.path.abspath, sorted(glob(path))))
+            if len(filenames) == 0:
+                raise FileNotFoundError(path)
         else:
             raise ValueError("'path' should be a string or a list of strings")
         self.files = [filetype(fn, *args, **kwargs) for fn in filenames]
         self.sizes = numpy.array([len(f) for f in self.files], dtype='i8')
 
         # set dtype and size
-        self.dtype = self.files[0].dtype
-        self.size  = self.sizes.sum()
+        FileType.__init__(self, dtype=self.files[0].dtype,
+                                size=self.sizes.sum())
 
     def __repr__(self):
         return "FileStack(%s, ... %d files)" % (repr(self.files[0]), self.nfiles)
@@ -97,7 +94,6 @@ class FileStack(FileType):
         data : array_like
             a numpy structured array holding the requested data
         """
-        if isinstance(columns, string_types): columns = [columns]
 
         toret = []
         for fnum in tools.get_file_slice(self.sizes, start, stop):